requirements.txt
fastapi
uvicorn
CouchDB == 1.2
PyYAML == 5.3.1